updated client

+++ NetTeam4/NetTeam4/src/Game/Player.h
@@ -5,7 +5,7 @@ class Player
 {
 public:
 	int Id;
-	int x, y, w, h;
+	float x, y, w, h;
 	int LastProcessedServerFrame;
 
 };
